==> vignettes/10_OM3_data_moderate_stock_assessment_main_report.Rmd <==
---
title: 'OM3 data-moderate stock assessments'
date: "`r Sys.Date()`"
output: 
  html_document:
    toc: TRUE
    number_sections: true
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, 
                      message = FALSE,
                      warning = FALSE,
                      fig.width = 12,
                      fig.height = 10,
                      results = 'asis')
options(scipen = 1, digits = 4)
```

## Just Another Bayesian Biomass Assessment (JABBA)

- Website: [https://github.com/jabbamodel/JABBA](https://github.com/jabbamodel/JABBA)
- Version: commit d73313debf9326f2be920846ee361b3dcca8c22d

```{r results='hide'}
devtools::load_all()

# Install required R packages -------------------------------------
if (!file.exists("~/.R/Makevars.win")){
  dir.create('~/.R')
  file.create('~/.R/Makevars')

  # Edit file and add JAGS_ROOT following instructions here:
  # https://cran.r-project.org/web/packages/rjags/INSTALL
  # JAGS_ROOT=c:/Progra~1/JAGS/JAGS-4.0.0
  file.edit("~/.R/Makevars.win")
}

required_pkg <- c(
  "devtools", "here", "gplots",
  "coda", "rjags", "R2jags",
  "fitdistrplus", "reshape",
  "ggplot2", "reshape2",
  "RColorBrewer", "scales", "fmsb",
  "gridExtra", "Polychrome"
)

pkg_to_install <- required_pkg[!(required_pkg %in%
                                   installed.packages()[, "Package"])]
if (length(pkg_to_install)) install.packages(pkg_to_install, repos = "http://cran.us.r-project.org")

lapply(required_pkg, library, character.only = TRUE)

devtools::install_github("debinqiu/snpar")
devtools::install_github("jabbamodel/JABBA",
                         ref = "d73313debf9326f2be920846ee361b3dcca8c22d"
)

library(JABBA)
```
# EwE without environmental effects and without fleet dynamics
<!-- ## Year 1996 Assessment with low CV in data and without environmental link in the estimation model -->

<!-- ```{r} -->
<!-- scenario_filename <- "_lowCV" -->
<!-- terminal_year <- 1996 -->
<!-- fishery_CV_input <- 0.05 -->
<!-- survey_CV_input <- 0.1 -->
<!-- add_environmental_effects <- FALSE -->
<!-- add_fleet_dynamics <- FALSE -->
<!-- cat(knitr::knit_child('_OM3_data_moderate_stock_assessment_template.Rmd', -->
<!--                       envir = environment(), -->
<!--                       quiet = TRUE)) -->
<!-- ``` -->

<!-- ## Year 1997 Assessment with low CV in data and without environmental link in the estimation model --> -->

<!-- ```{r} -->
<!-- scenario_filename <- "_lowCV" -->
<!-- terminal_year <- 1997 -->
<!-- fishery_CV_input <- 0.05 -->
<!-- survey_CV_input <- 0.1 -->
<!-- add_environmental_effects <- FALSE -->
<!-- add_fleet_dynamics <- FALSE -->
<!-- cat(knitr::knit_child('_OM3_data_moderate_stock_assessment_template.Rmd', -->
<!--                       envir = environment(), -->
<!--                       quiet = TRUE)) -->
<!-- ``` -->

<!-- ## Year 2004 Assessment with low CV in data and without environmental link in the estimation model -->

<!-- ```{r} -->
<!-- scenario_filename <- "_lowCV" -->
<!-- terminal_year <- 2004 -->
<!-- fishery_CV_input <- 0.05 -->
<!-- survey_CV_input <- 0.1 -->
<!-- add_environmental_effects <- FALSE -->
<!-- add_fleet_dynamics <- FALSE -->
<!-- cat(knitr::knit_child('_OM3_data_moderate_stock_assessment_template.Rmd', -->
<!--                       envir = environment(), -->
<!--                       quiet = TRUE)) -->
<!-- ``` -->

<!-- ## Year 2007 Assessment with low CV in data and without environmental link in the estimation model -->

<!-- ```{r} -->
<!-- scenario_filename <- "_lowCV" -->
<!-- terminal_year <- 2007 -->
<!-- fishery_CV_input <- 0.05 -->
<!-- survey_CV_input <- 0.1 -->
<!-- add_environmental_effects <- FALSE -->
<!-- add_fleet_dynamics <- FALSE -->
<!-- cat(knitr::knit_child('_OM3_data_moderate_stock_assessment_template.Rmd', -->
<!--                       envir = environment(), -->
<!--                       quiet = TRUE)) -->
<!-- ``` -->

## Year 2012 Assessment with low CV in data and without environmental link in the estimation model

```{r}
scenario_filename <- "_lowCV"
terminal_year <- 2012
fishery_CV_input <- 0.05
survey_CV_input <- 0.1
add_environmental_effects <- FALSE
add_fleet_dynamics <- FALSE
cat(knitr::knit_child('_OM3_data_moderate_stock_assessment_template.Rmd',
                      envir = environment(),
                      quiet = TRUE))
```

<!-- ## Year 1996 Assessment with high CV in data and without environmental link in the estimation model -->

<!-- ```{r} -->
<!-- scenario_filename <- "_highCV" -->
<!-- terminal_year <- 1996 -->
<!-- fishery_CV_input <- 0.2 -->
<!-- survey_CV_input <- 0.5 -->
<!-- add_environmental_effects <- FALSE -->
<!-- add_fleet_dynamics <- FALSE -->
<!-- cat(knitr::knit_child('_OM3_data_moderate_stock_assessment_template.Rmd', -->
<!--                       envir = environment(), -->
<!--                       quiet = TRUE)) -->
<!-- ``` -->

<!-- ## Year 1997 Assessment with high CV in data and without environmental link in the estimation model -->

<!-- ```{r} -->
<!-- scenario_filename <- "_highCV" -->
<!-- terminal_year <- 1997 -->
<!-- fishery_CV_input <- 0.2 -->
<!-- survey_CV_input <- 0.5 -->
<!-- add_environmental_effects <- FALSE -->
<!-- add_fleet_dynamics <- FALSE -->
<!-- cat(knitr::knit_child('_OM3_data_moderate_stock_assessment_template.Rmd', -->
<!--                       envir = environment(), -->
<!--                       quiet = TRUE)) -->
<!-- ``` -->

<!-- ## Year 2004 Assessment with high CV in data and without environmental link in the estimation model -->

<!-- ```{r} -->
<!-- scenario_filename <- "_highCV" -->
<!-- terminal_year <- 2004 -->
<!-- fishery_CV_input <- 0.2 -->
<!-- survey_CV_input <- 0.5 -->
<!-- add_environmental_effects <- FALSE -->
<!-- add_fleet_dynamics <- FALSE -->
<!-- cat(knitr::knit_child('_OM3_data_moderate_stock_assessment_template.Rmd', -->
<!--                       envir = environment(), -->
<!--                       quiet = TRUE)) -->
<!-- ``` -->

<!-- ## Year 2007 Assessment with high CV in data and without environmental link in the estimation model -->

<!-- ```{r} -->
<!-- scenario_filename <- "_highCV" -->
<!-- terminal_year <- 2007 -->
<!-- fishery_CV_input <- 0.2 -->
<!-- survey_CV_input <- 0.5 -->
<!-- add_environmental_effects <- FALSE -->
<!-- add_fleet_dynamics <- FALSE -->
<!-- cat(knitr::knit_child('_OM3_data_moderate_stock_assessment_template.Rmd', -->
<!--                       envir = environment(), -->
<!--                       quiet = TRUE)) -->
<!-- ``` -->

<!-- ## Year 2012 Assessment with high CV in data and without environmental link in the estimation model -->

<!-- ```{r} -->
<!-- scenario_filename <- "_highCV" -->
<!-- terminal_year <- 2012 -->
<!-- fishery_CV_input <- 0.2 -->
<!-- survey_CV_input <- 0.5 -->
<!-- add_environmental_effects <- FALSE -->
<!-- add_fleet_dynamics <- FALSE -->
<!-- cat(knitr::knit_child('_OM3_data_moderate_stock_assessment_template.Rmd', -->
<!--                       envir = environment(), -->
<!--                       quiet = TRUE)) -->
<!-- ``` -->

# EwE with environmental effects and without fleet dynamics
<!-- ## Year 1996 Assessment with low CV in data and without environmental link in the estimation model -->

<!-- ```{r} -->
<!-- scenario_filename <- "_lowCV" -->
<!-- terminal_year <- 1996 -->
<!-- fishery_CV_input <- 0.05 -->
<!-- survey_CV_input <- 0.1 -->
<!-- add_environmental_effects <- TRUE -->
<!-- add_fleet_dynamics <- FALSE -->
<!-- cat(knitr::knit_child('_OM3_data_moderate_stock_assessment_template.Rmd', -->
<!--                       envir = environment(), -->
<!--                       quiet = TRUE)) -->
<!-- ``` -->

<!-- ## Year 1997 Assessment with low CV in data and without environmental link in the estimation model -->

<!-- ```{r} -->
<!-- scenario_filename <- "_lowCV" -->
<!-- terminal_year <- 1997 -->
<!-- fishery_CV_input <- 0.05 -->
<!-- survey_CV_input <- 0.1 -->
<!-- add_environmental_effects <- TRUE -->
<!-- add_fleet_dynamics <- FALSE -->
<!-- cat(knitr::knit_child('_OM3_data_moderate_stock_assessment_template.Rmd', -->
<!--                       envir = environment(), -->
<!--                       quiet = TRUE)) -->
<!-- ``` -->

<!-- ## Year 2004 Assessment with low CV in data and without environmental link in the estimation model -->

<!-- ```{r} -->
<!-- scenario_filename <- "_lowCV" -->
<!-- terminal_year <- 2004 -->
<!-- fishery_CV_input <- 0.05 -->
<!-- survey_CV_input <- 0.1 -->
<!-- add_environmental_effects <- TRUE -->
<!-- add_fleet_dynamics <- FALSE -->
<!-- cat(knitr::knit_child('_OM3_data_moderate_stock_assessment_template.Rmd', -->
<!--                       envir = environment(), -->
<!--                       quiet = TRUE)) -->
<!-- ``` -->

<!-- ## Year 2007 Assessment with low CV in data and without environmental link in the estimation model -->

<!-- ```{r} -->
<!-- scenario_filename <- "_lowCV" -->
<!-- terminal_year <- 2007 -->
<!-- fishery_CV_input <- 0.05 -->
<!-- survey_CV_input <- 0.1 -->
<!-- add_environmental_effects <- TRUE -->
<!-- add_fleet_dynamics <- FALSE -->
<!-- cat(knitr::knit_child('_OM3_data_moderate_stock_assessment_template.Rmd', -->
<!--                       envir = environment(), -->
<!--                       quiet = TRUE)) -->
<!-- ``` -->

## Year 2012 Assessment with low CV in data and without environmental link in the estimation model

```{r}
scenario_filename <- "_lowCV"
terminal_year <- 2012
fishery_CV_input <- 0.05
survey_CV_input <- 0.1
add_environmental_effects <- TRUE
add_fleet_dynamics <- FALSE
cat(knitr::knit_child('_OM3_data_moderate_stock_assessment_template.Rmd',
                      envir = environment(),
                      quiet = TRUE))
```

<!-- ## Year 1996 Assessment with high CV in data and without environmental link in the estimation model -->

<!-- ```{r} -->
<!-- scenario_filename <- "_highCV" -->
<!-- terminal_year <- 1996 -->
<!-- fishery_CV_input <- 0.2 -->
<!-- survey_CV_input <- 0.5 -->
<!-- add_environmental_effects <- TRUE -->
<!-- add_fleet_dynamics <- FALSE -->
<!-- cat(knitr::knit_child('_OM3_data_moderate_stock_assessment_template.Rmd', -->
<!--                       envir = environment(), -->
<!--                       quiet = TRUE)) -->
<!-- ``` -->

<!-- ## Year 1997 Assessment with high CV in data and without environmental link in the estimation model -->

<!-- ```{r} -->
<!-- scenario_filename <- "_highCV" -->
<!-- terminal_year <- 1997 -->
<!-- fishery_CV_input <- 0.2 -->
<!-- survey_CV_input <- 0.5 -->
<!-- add_environmental_effects <- TRUE -->
<!-- add_fleet_dynamics <- FALSE -->
<!-- cat(knitr::knit_child('_OM3_data_moderate_stock_assessment_template.Rmd', -->
<!--                       envir = environment(), -->
<!--                       quiet = TRUE)) -->
<!-- ``` -->

<!-- ## Year 2004 Assessment with high CV in data and without environmental link in the estimation model -->

<!-- ```{r} -->
<!-- scenario_filename <- "_highCV" -->
<!-- terminal_year <- 2004 -->
<!-- fishery_CV_input <- 0.2 -->
<!-- survey_CV_input <- 0.5 -->
<!-- add_environmental_effects <- TRUE -->
<!-- add_fleet_dynamics <- FALSE -->
<!-- cat(knitr::knit_child('_OM3_data_moderate_stock_assessment_template.Rmd', -->
<!--                       envir = environment(), -->
<!--                       quiet = TRUE)) -->
<!-- ``` -->

<!-- ## Year 2007 Assessment with high CV in data and without environmental link in the estimation model -->

<!-- ```{r} -->
<!-- scenario_filename <- "_highCV" -->
<!-- terminal_year <- 2007 -->
<!-- fishery_CV_input <- 0.2 -->
<!-- survey_CV_input <- 0.5 -->
<!-- add_environmental_effects <- TRUE -->
<!-- add_fleet_dynamics <- FALSE -->
<!-- cat(knitr::knit_child('_OM3_data_moderate_stock_assessment_template.Rmd', -->
<!--                       envir = environment(), -->
<!--                       quiet = TRUE)) -->
<!-- ``` -->

<!-- ## Year 2012 Assessment with high CV in data and without environmental link in the estimation model -->

<!-- ```{r} -->
<!-- scenario_filename <- "_highCV" -->
<!-- terminal_year <- 2012 -->
<!-- fishery_CV_input <- 0.2 -->
<!-- survey_CV_input <- 0.5 -->
<!-- add_environmental_effects <- TRUE -->
<!-- add_fleet_dynamics <- FALSE -->
<!-- cat(knitr::knit_child('_OM3_data_moderate_stock_assessment_template.Rmd', -->
<!--                       envir = environment(), -->
<!--                       quiet = TRUE)) -->
<!-- ``` -->

# EwE with environmental effects and fleet dynamics
<!-- ## Year 1996 Assessment with low CV in data and without environmental link in the estimation model -->

<!-- ```{r} -->
<!-- terminal_year <- 1996 -->
<!-- fishery_CV_input <- 0.05 -->
<!-- survey_CV_input <- 0.1 -->
<!-- add_environmental_effects <- TRUE -->
<!-- add_fleet_dynamics <- TRUE -->
<!-- cat(knitr::knit_child('_OM3_data_moderate_stock_assessment_template.Rmd', -->
<!--                       envir = environment(), -->
<!--                       quiet = TRUE)) -->
<!-- ``` -->

<!-- ## Year 1997 Assessment with low CV in data and without environmental link in the estimation model -->

<!-- ```{r} -->
<!-- scenario_filename <- "_lowCV" -->
<!-- terminal_year <- 1997 -->
<!-- fishery_CV_input <- 0.05 -->
<!-- survey_CV_input <- 0.1 -->
<!-- add_environmental_effects <- TRUE -->
<!-- add_fleet_dynamics <- TRUE -->
<!-- cat(knitr::knit_child('_OM3_data_moderate_stock_assessment_template.Rmd', -->
<!--                       envir = environment(), -->
<!--                       quiet = TRUE)) -->
<!-- ``` -->

<!-- ## Year 2004 Assessment with low CV in data and without environmental link in the estimation model -->

<!-- ```{r} -->
<!-- scenario_filename <- "_lowCV" -->
<!-- terminal_year <- 2004 -->
<!-- fishery_CV_input <- 0.05 -->
<!-- survey_CV_input <- 0.1 -->
<!-- add_environmental_effects <- TRUE -->
<!-- add_fleet_dynamics <- TRUE -->
<!-- cat(knitr::knit_child('_OM3_data_moderate_stock_assessment_template.Rmd', -->
<!--                       envir = environment(), -->
<!--                       quiet = TRUE)) -->
<!-- ``` -->

<!-- ## Year 2007 Assessment with low CV in data and without environmental link in the estimation model -->

<!-- ```{r} -->
<!-- scenario_filename <- "_lowCV" -->
<!-- terminal_year <- 2007 -->
<!-- fishery_CV_input <- 0.05 -->
<!-- survey_CV_input <- 0.1 -->
<!-- add_environmental_effects <- TRUE -->
<!-- add_fleet_dynamics <- TRUE -->
<!-- cat(knitr::knit_child('_OM3_data_moderate_stock_assessment_template.Rmd', -->
<!--                       envir = environment(), -->
<!--                       quiet = TRUE)) -->
<!-- ``` -->

## Year 2012 Assessment with low CV in data and without environmental link in the estimation model

```{r}
scenario_filename <- "_lowCV"
terminal_year <- 2012
fishery_CV_input <- 0.05
survey_CV_input <- 0.1
add_environmental_effects <- TRUE
add_fleet_dynamics <- TRUE
cat(knitr::knit_child('_OM3_data_moderate_stock_assessment_template.Rmd',
                      envir = environment(),
                      quiet = TRUE))
```

<!-- ## Year 1996 Assessment with high CV in data and without environmental link in the estimation model -->

<!-- ```{r} -->
<!-- scenario_filename <- "_highCV" -->
<!-- terminal_year <- 1996 -->
<!-- fishery_CV_input <- 0.2 -->
<!-- survey_CV_input <- 0.5 -->
<!-- add_environmental_effects <- TRUE -->
<!-- add_fleet_dynamics <- TRUE -->
<!-- cat(knitr::knit_child('_OM3_data_moderate_stock_assessment_template.Rmd', -->
<!--                       envir = environment(), -->
<!--                       quiet = TRUE)) -->
<!-- ``` -->

<!-- ## Year 1997 Assessment with high CV in data and without environmental link in the estimation model -->

<!-- ```{r} -->
<!-- scenario_filename <- "_highCV" -->
<!-- terminal_year <- 1997 -->
<!-- fishery_CV_input <- 0.2 -->
<!-- survey_CV_input <- 0.5 -->
<!-- add_environmental_effects <- TRUE -->
<!-- add_fleet_dynamics <- TRUE -->
<!-- cat(knitr::knit_child('_OM3_data_moderate_stock_assessment_template.Rmd', -->
<!--                       envir = environment(), -->
<!--                       quiet = TRUE)) -->
<!-- ``` -->

<!-- ## Year 2004 Assessment with high CV in data and without environmental link in the estimation model -->

<!-- ```{r} -->
<!-- scenario_filename <- "_highCV" -->
<!-- terminal_year <- 2004 -->
<!-- fishery_CV_input <- 0.2 -->
<!-- survey_CV_input <- 0.5 -->
<!-- add_environmental_effects <- TRUE -->
<!-- add_fleet_dynamics <- TRUE -->
<!-- cat(knitr::knit_child('_OM3_data_moderate_stock_assessment_template.Rmd', -->
<!--                       envir = environment(), -->
<!--                       quiet = TRUE)) -->
<!-- ``` -->

<!-- ## Year 2007 Assessment with high CV in data and without environmental link in the estimation model -->

<!-- ```{r} -->
<!-- scenario_filename <- "_highCV" -->
<!-- terminal_year <- 2007 -->
<!-- fishery_CV_input <- 0.2 -->
<!-- survey_CV_input <- 0.5 -->
<!-- add_environmental_effects <- TRUE -->
<!-- add_fleet_dynamics <- TRUE -->
<!-- cat(knitr::knit_child('_OM3_data_moderate_stock_assessment_template.Rmd', -->
<!--                       envir = environment(), -->
<!--                       quiet = TRUE)) -->
<!-- ``` -->

<!-- ## Year 2012 Assessment with high CV in data and without environmental link in the estimation model -->

<!-- ```{r} -->
<!-- scenario_filename <- "_highCV" -->
<!-- terminal_year <- 2012 -->
<!-- fishery_CV_input <- 0.2 -->
<!-- survey_CV_input <- 0.5 -->
<!-- add_environmental_effects <- TRUE -->
<!-- add_fleet_dynamics <- TRUE -->
<!-- cat(knitr::knit_child('_OM3_data_moderate_stock_assessment_template.Rmd', -->
<!--                       envir = environment(), -->
<!--                       quiet = TRUE)) -->
<!-- ``` -->
