==> vignettes/16_OM3_Data_moderate_stock_assessmet.Rmd <==
---
title: 'OM3 data-moderate stock assessment'
date: "`r Sys.Date()`"
output:
  html_document: 
    code_folding: hide
  word_document: default
---
  
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
options(digits = 3)
```

## Just Another Bayesian Biomass Assessment (JABBA)

- Website: [https://github.com/jabbamodel/JABBA](https://github.com/jabbamodel/JABBA)
- Version: commit d73313debf9326f2be920846ee361b3dcca8c22d

```{r, results='hide', echo=FALSE, message=FALSE, warning=FALSE, fig.width=12, fig.height=10}
devtools::load_all()

# Install required R packages -------------------------------------
required_pkg <- c(
  "devtools", "here", "gplots",
  "coda", "rjags", "R2jags",
  "fitdistrplus", "reshape",
  "ggplot2", "reshape2",
  "RColorBrewer", "scales", "fmsb",
  "gridExtra"
)

pkg_to_install <- required_pkg[!(required_pkg %in%
                                   installed.packages()[, "Package"])]
if (length(pkg_to_install)) install.packages(pkg_to_install)

lapply(required_pkg, library, character.only = TRUE)

devtools::install_github("jabbamodel/JABBA",
                         ref = "d73313debf9326f2be920846ee361b3dcca8c22d"
)

library(JABBA)

```

## Case 0: stock assessment base run
```{r, results='hide', echo=FALSE, message=FALSE, warning=FALSE, fig.width=12, fig.height=10}
# Load simulated input data
source(here::here("Rscript", "simulationOM3.R"))# OM reference points ---------------------------------------------

# Reference points scenarios
# Maximum relative F = 7
# No. of steps = 700
# No. of trial years = 100
# Option 1: Search by fleets or groups
# Option 2: Full compensation or stationary system

functional_groups <- c(
  "StripedBass0",
  "StripedBass2_5",
  "StripedBass6",
  "AtlanticMenhaden0",
  "AtlanticMenhaden1",
  "AtlanticMenhaden2",
  "AtlanticMenhaden3",
  "AtlanticMenhaden4",
  "AtlanticMenhaden5",
  "AtlanticMenhaden6",
  "SpinyDogfish",
  "BluefishJuvenile",
  "BluefishAdult",
  "WeakfishJuvenile",
  "WeakfishAdult",
  "AtlanticHerring0_1",
  "AtlanticHerring2",
  "Anchovies",
  "Benthos",
  "Zooplankton",
  "Phytoplankton",
  "Detritus"
)
ages <- 0:6
start_year <- 1985
terminal_year <- 2006
projection_nyear <- 5
model_year <- start_year:terminal_year
projection_year <- (terminal_year+1):(terminal_year+projection_nyear)

file_path <- here::here("data", "ewe", "7ages_newsim_om3", "ewe7ages_ecosim_om3", "msy_forcing_pdsi_egg_amo1", "fleet")

compensation_msy <- read_ewe_reference_points(
  file_path = file_path,
  file_names = c(
    "F.menhaden_Catch_FullCompensation.csv",
    "F.menhaden_B_FullCompensation.csv"
  ),
  reference_points_scenario = "compensation",
  functional_groups = functional_groups,
  key_functional_group = "AtlanticMenhaden",
  ages = ages,
  plot = TRUE
) # MSY, FMSY, BMSY: 0.5913518, 2.939998, 1.619142

stationary_msy <- read_ewe_reference_points(
  file_path = file_path,
  file_names = c(
    "F.menhaden_Catch_StationarySystem.csv",
    "F.menhaden_B_StationarySystem.csv"
  ),
  reference_points_scenario = "stationary",
  functional_groups = functional_groups,
  key_functional_group = "AtlanticMenhaden",
  ages = ages,
  plot = TRUE
) # MSY, FMSY, BMSY: 0.5784879, 2.899998, 1.604036

msy_mean <- sapply(1:length(compensation_msy), function(x) {mean(c(compensation_msy[[x]], stationary_msy[[x]]))})
names(msy_mean) <- names(compensation_msy)

# OM unfished biomass --------------------------------------------
age_name <- paste0("AtlanticMenhaden", ages)

# Only menhaden F = 0, other functional groups: F = stock assessment F from 1987 - 2017, then F = 0 for the rest of the years
biomass <- read_ewe_output(
  file_path = here::here("data", "ewe", "7ages_newsim_om3", "ewe7ages_ecosim_om3_F0_menhaden", "ecosim_forcing_pdsi_egg_amo1"),
  file_names = "biomass_monthly.csv",
  skip_nrows = 8,
  plot = FALSE,
  figure_titles = NULL,
  functional_groups = functional_groups,
  figure_colors = NULL
)
time_id <- seq(1, nrow(biomass[[1]]), by = 12)
biomass_f0_menhaden <- apply(biomass[[1]][, age_name], 1, sum)[time_id] * 1000000

# All functional groups F = 0
biomass <- read_ewe_output(
  file_path = here::here("data", "ewe", "7ages_newsim_om3", "ewe7ages_ecosim_om3_F0_all_functional_groups", "ecosim_forcing_pdsi_egg_amo1"),
  file_names = "biomass_monthly.csv",
  skip_nrows = 8,
  plot = FALSE,
  figure_titles = NULL,
  functional_groups = functional_groups,
  figure_colors = NULL
)
time_id <- seq(1, nrow(biomass[[1]]), by = 12)
biomass_f0_all <- apply(biomass[[1]][, age_name], 1, sum)[time_id] * 1000000

x <- start_year:(start_year + length(time_id) - 1)
yrange <- range(biomass_f0_menhaden, biomass_f0_all)
plot(x, biomass_f0_menhaden,
     xlab = "Year", ylab = "Biomass",
     lty = 1, type = "l"
)
lines(x, biomass_f0_all,
      pch = 1, lty = 2
)
legend("topright",
       c("Menhaden F = 0", "All functional groups F = 0"),
       lty = c(1, 2),
       bty = "n"
)

B0_F0_menhaden <- mean(tail(biomass_f0_menhaden, n = length(ages) * 2))
B0_F0_all <- mean(tail(biomass_f0_all, n = length(ages) * 2))
B0_F0_mean <- mean(c(B0_F0_menhaden, B0_F0_all))

# Load OM data
file_path <- here::here("data", "ewe", "7ages_newsim_om3", "ewe7ages_ecosim_om3", "ecosim_forcing_pdsi_egg_amo1")
source(here::here("Rscript", "load_indices.R"))
set.seed(999)
# Load EwE biomass data

age_name <- paste0("AtlanticMenhaden", ages)

biomass <- read_ewe_output(
  file_path = file_path,
  file_names = "biomass_monthly.csv",
  skip_nrows = 8,
  plot = FALSE,
  figure_titles = NULL,
  functional_groups = functional_groups,
  figure_colors = NULL
)

time_id <- seq(1, nrow(biomass[[1]]), by = 12)[1:(length(model_year)+length(projection_year))]
biomass_ewe <- apply(biomass[[1]][, age_name], 1, sum) * 1000000
# Set up JABBA base case assessment

output_dir <- here::here("data", "data_moderate", "om3")
if (!dir.exists(output_dir)) dir.create(output_dir)
compensation_msy$BMSY/B0_F0_menhaden
# case 0
# BmsyK = 0.4 # shape = 1.2
# BmsyK = 0.45 # shape = 1.5
BmsyK = 0.5 # shape = 2
# BmsyK = , # shape = 3.23
# BmsyK = msy_mean["BMSY"]/B0_F0_mean, # shape = 7.6
## Do not use effort data
ss_case0_input_noEffort <- generate_jabba(
  assessment_name = "case0_noEffort",
  output_dir = output_dir,
  sa_data = sa_data,
  BmsyK = BmsyK,
  model_year = model_year,
  projection_year = projection_year,
  effort_data = NULL
)


# k.prior (mu.k, cv.k)
# k_init <- B0_F0_menhaden
K.init <- 8*max(sa_data$fishery$obs_total_catch_biomass$fleet1)
K.prior = c(K.init, max(sa_data$fishery$obs_total_catch_biomass$fleet1))
log.K =  mean(log(K.prior))
sd.K= abs(log.K - log(K.prior[1]))/2
log.K = mean(log(K.prior))+0.5*sd.K^2 # Will be back-bias corrected in lognorm function
CV.K = sqrt(exp(sd.K^2)-1) #
# r.prior (mu.r, sd.r) # mu.r range: 0.2 - 0.8
start.r <- c(0.2,0.8)
log.r = mean(log(start.r))
sd.r = abs(log.r - log(start.r[1]))/2


# ss_case0_input_noEffort$jagsdata$K.pr <- c(k_init, 0.1)
# ss_case0_input_noEffort$jagsdata$r.pr <- c(0.2, 0.1)
# # ss_case0_input_noEffort$jagsdata$r.pr <- c(0.6, 0.1)
# psi_init <- biomass_ewe[1]/k_init
# ss_case0_input_noEffort$jagsdata$psi.pr <- c(psi_init, 0.1)

# k_init <- B0_F0_all
# k_init <- B0_F0_menhaden
ss_case0_input_noEffort$jagsdata$K.pr <- c(exp(log.K), CV.K) # 5208722, 2091
# ss_case0_input_noEffort$jagsdata$K.pr <- c(k_init, 0.5)
ss_case0_input_noEffort$jagsdata$r.pr <- c(exp(log.r), sd.r) # 0.4, 0.35
# psi_init <- biomass_ewe[1]/k_init
psi_init <- biomass_ewe[1]/exp(log.K)
ss_case0_input_noEffort$jagsdata$psi.pr <- c(psi_init, 0.25)


ss_case0_noEffort <- JABBA::fit_jabba(
  jbinput = ss_case0_input_noEffort,
  save.jabba = TRUE,
  save.all = TRUE,
  save.prj = TRUE,
  output.dir = output_dir,
  save.csvs = T
)

ss_case0_noEffort$estimates
write.csv(ss_case0_noEffort$estimates, file.path(output_dir, "case0_noEffort_estimates.csv"))


## Use effort data
year_id <- seq(1, nrow(amo_unsmooth_lag1), by = 12)
effort <- apply(ewe_faa[, grep("menhaden", colnames(ewe_faa))] / ewe_catchability[year_id, grep("menhaden", colnames(ewe_catchability))], 1, sum)

ss_case0_effort_input <- generate_jabba(
  assessment_name = "case0",
  output_dir = output_dir,
  sa_data = sa_data,
  BmsyK = BmsyK,
  model_year = model_year,
  projection_year = projection_year,
  effort_data = effort[1:length(model_year)]
)

# k_init <- max(sa_data$fishery$obs_total_catch_biomass$fleet1) * 10
# k_init <- 3500000
# k_init <- B0_F0_menhaden
# ss_case0_effort_input$jagsdata$K.pr <- c(k_init, 0.1)
# ss_case0_effort_input$jagsdata$r.pr <- c(0.2, 0.1)
# # ss_case0_effort_input$jagsdata$r.pr <- c(0.6, 0.1)
# psi_init <- biomass_ewe[1]/k_init
# ss_case0_effort_input$jagsdata$psi.pr <- c(psi_init, 0.1)

# k_init <- B0_F0_all
# k_init <- B0_F0_menhaden
ss_case0_effort_input$jagsdata$K.pr <- c(exp(log.K), CV.K) # 5208722, 2091
# ss_case0_input_noEffort$jagsdata$K.pr <- c(k_init, 0.5)
ss_case0_effort_input$jagsdata$r.pr <- c(exp(log.r), sd.r) # 0.4, 0.35
# psi_init <- biomass_ewe[1]/k_init
psi_init <- biomass_ewe[1]/exp(log.K)
ss_case0_effort_input$jagsdata$psi.pr <- c(psi_init, 0.25)

set.seed(999)
ss_case0_effort <- JABBA::fit_jabba(
  jbinput = ss_case0_effort_input,
  save.jabba = TRUE,
  save.all = TRUE,
  save.prj = TRUE,
  output.dir = output_dir,
  save.csvs = T
)

ss_case0_effort$estimates
write.csv(ss_case0_effort$estimates, file.path(output_dir, "case0_estimates.csv"))

ss_case0_input <-  ss_case0_effort_input
ss_case0 <- ss_case0_effort

# Plot figures using JABBA
figure_dir <- file.path(output_dir, "noEffort")
if (!dir.exists(figure_dir)) dir.create(figure_dir)
JABBA::jabba_plots(jabba = ss_case0_noEffort, output.dir = figure_dir)
# plot with CIs (80% for projections)
# JABBA::jbplot_prj(ss_case0_noEffort, type = "BBmsy", output.dir = figure_dir)
# JABBA::jbplot_prj(ss_case0_noEffort, type = "BB0", output.dir = figure_dir)
# JABBA::jbplot_prj(ss_case0_noEffort, type = "FFmsy", output.dir = figure_dir)

figure_dir <- file.path(output_dir, "Effort")
if (!dir.exists(figure_dir)) dir.create(figure_dir)
JABBA::jabba_plots(jabba = ss_case0_effort, output.dir = figure_dir)
# plot with CIs (80% for projections)
# JABBA::jbplot_prj(ss_case0_effort, type = "BBmsy", output.dir = figure_dir)
# JABBA::jbplot_prj(ss_case0_effort, type = "BB0", output.dir = figure_dir)
# JABBA::jbplot_prj(ss_case0_effort, type = "FFmsy", output.dir = figure_dir)

# Plot figures for cases 0.1 and 0.2 ----------------------------------------------------
fmsy0 <- rep(ss_case0_noEffort$posteriors$Hmsy, times = (length(projection_year) + 1) * length(ss_case0_input_noEffort$jagsdata$TAC[1, ]))
ss_case0_noEffort$prj_posterior$f <- ss_case0_noEffort$prj_posterior$harvest * fmsy0

bmsy0 <- rep(ss_case0_noEffort$posteriors$SBmsy, times = (length(projection_year) + 1) * length(ss_case0_input_noEffort$jagsdata$TAC[1, ]))
ss_case0_noEffort$prj_posterior$biomass <- ss_case0_noEffort$prj_posterior$stock * bmsy0

figure_path <- here::here("figure", "data_moderate")
if (!dir.exists(figure_path)) dir.create(figure_path)

jpeg(filename = file.path(figure_path, paste0("Biomass", terminal_year, ".jpeg")), width = 200, height = 120, units = "mm", res = 1800)
# plot biomass over time
par(mfrow = c(1, 2))

ylim <- range(
  biomass_ewe[time_id],

  ss_case0_noEffort$timeseries[, , "B"],
  ss_case0_noEffort$prj_posterior$biomass[(round(ss_case0_noEffort$prj_posterior$tac) %in% round(mean(tail(ss_case0_input_noEffort$data$catch$Total, n = 3))))],

  ss_case0$timeseries[, , "B"],
  ss_case0$prj_posterior$biomass[(round(ss_case0$prj_posterior$tac) %in% round(mean(tail(ss_case0_input$data$catch$Total, n = 3))))]
)

# Plot biomass over time based on no effort model
plot(c(model_year, projection_year),
     biomass_ewe[time_id],
     xlab = "Year", ylab = "Biomass (mt)",
     ylim = ylim,
     pch = 16
)

lines(model_year, ss_case0_noEffort$timeseries[, "mu", "B"])
lines(model_year, ss_case0_noEffort$timeseries[, "lci", "B"], lty = 2)
lines(model_year, ss_case0_noEffort$timeseries[, "uci", "B"], lty = 2)


# Plot biomass based on last three years of catch

for (i in 1:length(projection_year)) {
  # Projection TAC input used mean catch from the recent 3 years
  subdata_id <- (ss_case0_noEffort$prj_posterior$year == projection_year[i]) & (round(ss_case0_noEffort$prj_posterior$tac) %in% round(mean(tail(ss_case0_input_noEffort$data$catch$Total, n = 3))))

  boxplot(ss_case0_noEffort$prj_posterior$biomass[subdata_id],
          add = TRUE,
          at = projection_year[i],
          col = "gray",
          width = 1,
          outline = TRUE,
          axes = FALSE
  )
}
points(c(model_year, projection_year),
       biomass_ewe[time_id],
       pch=16)
legend("topleft", "No CPUE index", bty = "n")

# Plot biomass over time based model that is using effort data
fmsy0 <- rep(ss_case0$posteriors$Hmsy, times = (length(projection_year) + 1) * length(ss_case0_input$jagsdata$TAC[1, ]))
ss_case0$prj_posterior$f <- ss_case0$prj_posterior$harvest * fmsy0

bmsy0 <- rep(ss_case0$posteriors$SBmsy, times = (length(projection_year) + 1) * length(ss_case0_input$jagsdata$TAC[1, ]))
ss_case0$prj_posterior$biomass <- ss_case0$prj_posterior$stock * bmsy0

plot(c(model_year, projection_year),
     biomass_ewe[time_id],
     xlab = "Year", ylab = "Biomass (mt)",
     ylim = ylim,
     pch = 16
)

lines(model_year, ss_case0$timeseries[, "mu", "B"])
lines(model_year, ss_case0$timeseries[, "lci", "B"], lty = 2)
lines(model_year, ss_case0$timeseries[, "uci", "B"], lty = 2)

cor(biomass_ewe[time_id][1:length(model_year)], ss_case0$timeseries[, "mu", "B"])

# Plot biomass based on last three years of catch
for (i in 1:length(projection_year)) {

  subdata_id <- (ss_case0$prj_posterior$year == projection_year[i]) & (round(ss_case0$prj_posterior$tac) %in% round(mean(tail(ss_case0_input$data$catch$Total, n = 3))))

  boxplot(ss_case0$prj_posterior$biomass[subdata_id],
          add = TRUE,
          at = projection_year[i],
          col = "gray",
          width = 1,
          outline = TRUE,
          axes = FALSE
  )
}
points(c(model_year, projection_year),
       biomass_ewe[time_id],
       pch=16)
legend("topleft", "CPUE index included", bty = "n")
dev.off()
# Plot mortality over time

mortality <- read.csv(file = here::here("data", "ewe", "7ages_newsim_om3", "fatage.csv"))
mortality_ewe_apical <- apply(mortality[, 2:ncol(mortality)], 1, max)
mortality_ewe_average <- apply(mortality[, 2:ncol(mortality)], 1, mean)

jpeg(filename = file.path(figure_path, paste0("fishing_mortality", terminal_year, ".jpeg")), width = 200, height = 120, units = "mm", res = 1800)
par(mfrow = c(1, 1))
ylim <- range(mortality_ewe_apical,
              mortality_ewe_average,
              ss_case0$timeseries[, "mu", "F"])
plot(c(model_year, projection_year),
     #mortality_ewe[time_id],
     mortality_ewe_apical[1:(length(model_year)+projection_nyear)],
     type = "o",
     xlab = "Year", ylab = "F",
     pch = 1,
     col = 1,
     lty = 1,
     ylim = ylim
)
lines(c(model_year, projection_year),
      mortality_ewe_average[1:(length(model_year)+projection_nyear)],
      pch = 2, type = "o", lty = 2, col = 2)
lines(model_year, ss_case0$timeseries[, "mu", "F"], lty = 3, col =3)
legend("topleft",
       c("EwE Apical F", "EwE Average F", "JABBA F"),
       bty = "n",
       pch = c(1, 2, NA),
       lty = 1:3,
       col = 1:3
)

cor(mortality_ewe_average[1:length(model_year)], ss_case0$timeseries[, "mu", "F"])
dev.off()
```


## Cases 1-5 are based on the settings from case 0
- Case 1: Link Atlantic Multidecadal Oscillation Index with menhaden biomass estimates and adjust projections: AMO is an indicator of climate conditions and would affect recruitment variability of menhaden-like species
- Case 2: Link Palmer drought severity index with menhaden biomass estimates and adjust projections: PSDI is a long-term indicator of drought conditions and it reflects river discharge and precipitation
- Case 3: Link biomass of Striped bass from the EwE with menhaden biomass estimates and adjust projections because bass is a major predator
- Case 4: Link fishing effort of menhaden with menhaden biomass estimates and adjust projections
- Case 5: Link catch per unit effort of menhaden with menhaden biomass estimates and adjust projections
- Linear regression models from case 1 - 4
  - True biomass of menhaden-like species as functions of AMO, PDSI, biomass of striped bass, and fishing effort of menhaden

### Status of indicators (SOI)

- If stock-indicator relationship is positive, $SOI_{y} = \frac{I_{y}-I_{y}^{min}}{I_{y}^{max}-I_{y}^{min}}$
  
- If stock-indicator relationship is negative, $SOI_{y} = 1-\frac{I_{y}-I_{y}^{min}}{I_{y}^{max}-I_{y}^{min}}$
  
  where $I_{y}$ represents indicator value in year y. $I_{y}^{min}$ and $I_{y}^{max}$ represent the minimum and maximum values of $I$ from the time series.

```{r,  echo=FALSE, message=FALSE, warning=FALSE, fig.width=12, fig.height=10}

lm_slope <- data.frame(
  case = "True indices",
  amo = NA,
  #pcp = NA,
  bassB = NA,
  #price = NA,
  pdsi = NA,
  effort = NA,
  cpue = NA
)

year_id <- seq(1, nrow(amo_unsmooth_lag1), by = 12)
index_year <- c(model_year)
lag = 1
biomass_lag_id <- (1+lag):length(index_year)
index_lag_id <- 1:(length(index_year)-lag)

amo <- amo_unsmooth_lag1[year_id, ]
amo_lm <- lm(biomass_ewe[time_id][biomass_lag_id] ~ amo$scaled_value[index_lag_id])
amo_fit <- fitted(amo_lm)
lm_slope$amo <- paste0(
  round(summary(amo_lm)$coefficients[2, 1], digits = 2),
  if (summary(amo_lm)$coefficients[2, 4] <= 0.05) {
    "*"
  }
)

# pcp <- precipitation[year_id, ]
# pcp_lm <- lm(biomass_ewe[time_id][biomass_lag_id] ~ pcp$scaled_value[index_lag_id])
# pcp_fit <- fitted(pcp_lm)
# lm_slope$pcp <- paste0(
#   round(summary(pcp_lm)$coefficients[2, 1], digits = 2),
#   if(summary(pcp_lm)$coefficients[2, 4] <= 0.05) {"*"})

pdsi <- palmer_drought_severity_index[year_id, ]
pdsi_lm <- lm(biomass_ewe[time_id][biomass_lag_id] ~ pdsi$scaled_value[index_lag_id])
pdsi_fit <- fitted(pdsi_lm)
lm_slope$pdsi <- paste0(
  round(summary(pdsi_lm)$coefficients[2, 1], digits = 2),
  if(summary(pdsi_lm)$coefficients[2, 4] <= 0.05) {"*"})

bassB <- bass_bio[bass_bio$Year %in% year_id, ]
bassB_lm <- lm(biomass_ewe[time_id][biomass_lag_id] ~ bassB$bass_bio[index_lag_id])
bassB_fit <- fitted(bassB_lm)
lm_slope$bassB <- paste0(
  round(summary(bassB_lm)$coefficients[2, 1], digits = 2),
  if(summary(bassB_lm)$coefficients[2, 4] <= 0.05) {"*"})

# sub_menhadenP <- menhaden_price$Inflation.Adjust.Price.Per.MT[menhaden_price$Year %in% index_year]
# price_lm <- lm(biomass_ewe[time_id][biomass_lag_id] ~ sub_menhadenP[index_lag_id])
# price_fit <- fitted(price_lm)
# lm_slope$price <- paste0(
#   round(summary(price_lm)$coefficients[2, 1], digits = 2),
#   if(summary(price_lm)$coefficients[2, 4] <= 0.05) {"*"})

effort <- apply(ewe_faa[, grep("menhaden", colnames(ewe_faa))] / ewe_catchability[year_id, grep("menhaden", colnames(ewe_catchability))], 1, sum)
effort_lm <- lm(biomass_ewe[time_id][biomass_lag_id] ~ effort[index_lag_id])
effort_fit <- fitted(effort_lm)
lm_slope$effort <- paste0(
  round(summary(effort_lm)$coefficients[2, 1], digits = 2),
  if(summary(effort_lm)$coefficients[2, 4] <= 0.05) {"*"})

cpue <- sa_data$fishery$obs_total_catch_biomass$fleet1/effort
cpue_lm <- lm(biomass_ewe[time_id][biomass_lag_id] ~ cpue[index_lag_id])
cpue_fit <- fitted(cpue_lm)
lm_slope$cpue <- paste0(
  round(summary(cpue_lm)$coefficients[2, 1], digits = 2),
  if(summary(cpue_lm)$coefficients[2, 4] <= 0.05) {"*"})

# linear regression analysis summary table
lm_data_om <- rbind(
  data.frame(
    Menhaden_Biomass = biomass_ewe[time_id][biomass_lag_id],
    Variable = "AMO",
    Index_Value = amo$scaled_value[index_lag_id]
  ),
  data.frame(
    Menhaden_Biomass = biomass_ewe[time_id][biomass_lag_id],
    Variable = "PDSI",
    Index_Value = pdsi$scaled_value[index_lag_id]
  ),
  data.frame(
    Menhaden_Biomass = biomass_ewe[time_id][biomass_lag_id],
    Variable = "Biomass of Striped bass",
    Index_Value = bassB$bass_bio[index_lag_id]
  ),
  data.frame(
    Menhaden_Biomass = biomass_ewe[time_id][biomass_lag_id],
    Variable = "Menhaden fishing effort",
    Index_Value = effort[index_lag_id]
  ),
  data.frame(
    Menhaden_Biomass = biomass_ewe[time_id][biomass_lag_id],
    Variable = "Menhaden CPUE",
    Index_Value = cpue[index_lag_id]
  )
)

lm_data_om$model <- "OM"

jpeg(filename = file.path(figure_path, paste0("lm_slope_om_", terminal_year, ".jpeg")), width = 200, height = 100, units = "mm", res = 1800)
p <- gridExtra::tableGrob(lm_slope)
gridExtra::grid.arrange(p)
dev.off()

# Linear regression analysis figure
par(mfrow = c(3, 2))

plot(amo$scaled_value[index_lag_id], biomass_ewe[time_id][biomass_lag_id],
     xlab = "AMO values",
     ylab = "Biomass of menhaden-like species from OM"
)
abline(amo_lm)

# plot(pcp$scaled_value[index_lag_id], biomass_ewe[time_id][biomass_lag_id],
#      xlab = "Precipitation values",
#      ylab = "Biomass of menhaden-like species from OM"
# )
# abline(pcp_lm)

plot(pdsi$scaled_value[index_lag_id], biomass_ewe[time_id][biomass_lag_id],
     xlab = "PDSI values",
     ylab = "Biomass of menhaden-like species from OM"
)
abline(pdsi_lm)

plot(bassB$bass_bio[index_lag_id], biomass_ewe[time_id][biomass_lag_id],
     xlab = "Biomass of Striped bass",
     ylab = "Biomass of menhaden-like species from OM"
)
abline(bassB_lm)

# plot(sub_menhadenP[index_lag_id], biomass_ewe[time_id][biomass_lag_id],
#      xlab = "Menhaden price",
#      ylab = "Biomass of menhaden-like species from OM"
# )
# abline(price_lm)

plot(effort[index_lag_id], biomass_ewe[time_id][biomass_lag_id],
     xlab = "Fishing effort of menhaden-like species from OM",
     ylab = "Biomass of menhaden-like species from OM"
)
abline(effort_lm)

plot(cpue[index_lag_id], biomass_ewe[time_id][biomass_lag_id],
     xlab = "CPUE of menhaden-like species from OM",
     ylab = "Biomass of menhaden-like species from OM"
)
abline(cpue_lm)

```

### Adjust projections

- If $\frac{B2012_{i}}{BMSY} > 1$, $F^{'}_{i} = FMSY^{min} + SOI2012 \times (FMSY^{max}-FMSY^{min})$

- If $\frac{B2012_{i}}{BMSY} \le 1$ and $\frac{B2012_{i}}{BMSY} > 0.5$, $F^{'}_{i} = SOI2012 \times \frac{B2012_{i}}{BMSY} \times FMSY_{i}$
  
- If $\frac{B2012_{i}}{BMSY} \le 0.5$, $F^{'}_{i} = 0$

where $i$ represents individual iterations


```{r, results='hide', echo=FALSE, message=FALSE, warning=FALSE, fig.width=12, fig.height=10}

# Projection: cases 1 - 5 ---------------------------

lm_slope <- data.frame(
  case = "PDSI+AMO1 Case",
  projection_year = 1:length(projection_year),
  amo = NA,
  pdsi = NA,
  bassB = NA,
  effort = NA, 
  cpue = NA
)

for (projection_year_id in 1:length(projection_year)){
  
  menhaden_b <- ss_case0$timeseries[, "mu", "B"]
  
  year_id <- seq(1, nrow(amo_unsmooth_lag1), by = 12)[1:length(model_year)]
  
  index_year = model_year
  
  
  # Linear regression model -----------------------------------------
  
  lag = 1
  biomass_lag_id <- (1+lag):length(index_year)
  index_lag_id <- 1:(length(index_year)-lag)
  
  amo <- amo_unsmooth_lag1[year_id, ]
  amo_lm <- lm(menhaden_b[biomass_lag_id] ~ amo$scaled_value[index_lag_id])
  amo_fit <- fitted(amo_lm)
  lm_slope$amo <- paste0(
    round(summary(amo_lm)$coefficients[2, 1], digits = 2),
    if (summary(amo_lm)$coefficients[2, 4] <= 0.05) {
      "*"
    }
  )
  
  pdsi <- palmer_drought_severity_index[year_id, ]
  pdsi_lm <- lm(menhaden_b[biomass_lag_id] ~ pdsi$scaled_value[index_lag_id])
  pdsi_fit <- fitted(pdsi_lm)
  lm_slope$pdsi <- paste0(
    round(summary(pdsi_lm)$coefficients[2, 1], digits = 2),
    if (summary(pdsi_lm)$coefficients[2, 4] <= 0.05) {
      "*"
    }
  )
  
  bassB <- bass_bio[bass_bio$Year %in% year_id, ]
  bassB_lm <- lm(menhaden_b[biomass_lag_id] ~ bassB$bass_bio[index_lag_id])
  bassB_fit <- fitted(bassB_lm)
  lm_slope$bassB <- paste0(
    round(summary(bassB_lm)$coefficients[2, 1], digits = 2),
    if (summary(bassB_lm)$coefficients[2, 4] <= 0.05) {
      "*"
    }
  )
  
  effort <- apply(ewe_faa[ewe_faa$X %in% ewe_faa$X[1:length(year_id)] , grep("menhaden", colnames(ewe_faa))] / ewe_catchability[year_id, grep("menhaden", colnames(ewe_catchability))], 1, sum)
  effort_lm <- lm(menhaden_b[biomass_lag_id] ~ effort[index_lag_id])
  effort_fit <- fitted(effort_lm)
  lm_slope$effort <- paste0(
    round(summary(effort_lm)$coefficients[2, 1], digits = 2),
    if (summary(effort_lm)$coefficients[2, 4] <= 0.05) {
      "*"
    }
  )
  
  cpue <- (sa_data$fishery$om_total_catch_biomass/effort)[1:length(year_id)]
  cpue_lm <- lm(menhaden_b[biomass_lag_id] ~ cpue[index_lag_id])
  cpue_fit <- fitted(cpue_lm)
  lm_slope$cpue <- paste0(
    round(summary(cpue_lm)$coefficients[2, 1], digits = 2),
    if (summary(cpue_lm)$coefficients[2, 4] <= 0.05) {
      "*"
    }
  )
  
  if (projection_year_id == 1) {
    lm_data_em <- rbind(
      data.frame(
        Menhaden_Biomass = menhaden_b[biomass_lag_id],
        Variable = "AMO",
        Index_Value = amo$scaled_value[index_lag_id]
      ),
      data.frame(
        Menhaden_Biomass = menhaden_b[biomass_lag_id],
        Variable = "PDSI",
        Index_Value = pdsi$scaled_value[index_lag_id]
      ),
      data.frame(
        Menhaden_Biomass = menhaden_b[biomass_lag_id],
        Variable = "Biomass of Striped bass",
        Index_Value = bassB$bass_bio[index_lag_id]
      ),
      data.frame(
        Menhaden_Biomass = menhaden_b[biomass_lag_id],
        Variable = "Menhaden fishing effort",
        Index_Value = effort[index_lag_id]
      ),
      data.frame(
        Menhaden_Biomass = menhaden_b[biomass_lag_id],
        Variable = "Menhaden CPUE",
        Index_Value = cpue[index_lag_id]
      )
    )
    lm_data_em$model <- "EM"
  }
  
  # if (projection_year_id == length(projection_year)){
  #   
  #   par(mfrow = c(2, 2))
  #   
  #   plot(amo$scaled_value[index_lag_id], menhaden_b[biomass_lag_id],
  #        xlab = "AMO values",
  #        ylab = "Biomass of menhaden-like species"
  #   )
  #   abline(amo_lm)
  #   
  #   plot(pdsi$scaled_value[index_lag_id], menhaden_b[biomass_lag_id],
  #        xlab = "PDSI values",
  #        ylab = "Biomass of menhaden-like species"
  #   )
  #   abline(pdsi_lm)
  #   
  #   plot(bassB$bass_bio[index_lag_id], menhaden_b[biomass_lag_id],
  #        xlab = "Biomass of Striped bass",
  #        ylab = "Biomass of menhaden-like species"
  #   )
  #   abline(bassB_lm)
  #   
  #   plot(effort[index_lag_id], menhaden_b[biomass_lag_id],
  #        xlab = "Menhaden fishing effort",
  #        ylab = "Biomass of menhaden-like species"
  #   )
  #   abline(effort_lm)
  #   
  # }
  
  
  # status of indicators --------------------------------------------
  
  amo_soi <- calc_soi(
    indicator_data = amo$scaled_value,
    slope = coef(amo_lm)[2]
  )
  
  pdsi_soi <- calc_soi(
    indicator_data = pdsi$scaled_value,
    slope = coef(pdsi_lm)[2]
  )
  
  bassB_soi <- calc_soi(
    indicator_data = bassB$bass_bio,
    slope = coef(bassB_lm)[2]
  )
  
  effort_soi <- calc_soi(
    indicator_data = effort,
    slope = coef(effort_lm)[2]
  )
  
  cpue_soi <- calc_soi(
    indicator_data = cpue,
    slope = coef(cpue_lm)[2]
  )
  
  if (projection_year_id == 1) {
    scaled_data <- data.frame(
      year = model_year,
      projection_year_id = projection_year_id,
      amo = scale(amo$scaled_value)[, 1],
      pdsi = scale(pdsi$scaled_value)[, 1],
      bassB = scale(bassB$bass_bio)[, 1],
      effort = scale(effort)[, 1],
      cpue = scale(cpue)[,1],
      menhadenB = scale(menhaden_b)[, 1]
    )
    
    soi_data <- data.frame(
      year = model_year,
      projection_year_id = projection_year_id,
      amo = amo_soi,
      pdsi = pdsi_soi,
      bass_b = bassB_soi,
      effort = effort_soi, 
      cpue = cpue_soi
    )
  } else {
    scaled_data <- rbind(
      scaled_data,
      data.frame(
        year = index_year,
        projection_year_id = projection_year_id,
        amo = scale(amo$scaled_value)[, 1],
        pdsi = scale(pdsi$scaled_value)[, 1],
        bassB = scale(bassB$bass_bio)[, 1],
        effort = scale(effort)[, 1],
        cpue = scale(cpue)[, 1],
        menhadenB = scale(menhaden_b)[, 1]
      )
    )
    
    soi_data <- rbind(
      soi_data,
      data.frame(
        year = index_year,
        projection_year_id = projection_year_id,
        amo = amo_soi,
        pdsi = pdsi_soi,
        bass_b = bassB_soi,
        effort = effort_soi,
        cpue = cpue_soi
      )
    )
  }
  
  
  # Adjusted FMSY ----------------------------------------------------
  Bt_BMSY <- ss_case0$posteriors$BtoBmsy[, length(model_year)]
  
  amo_fmsy <- adjust_projection_jabba(
    FMSY = ss_case0$refpts_posterior$Fmsy,
    soi = tail(amo_soi, n = 1),
    Bt_BMSY = as.matrix(Bt_BMSY)
  )
  pdsi_fmsy <- adjust_projection_jabba(
    FMSY = ss_case0$refpts_posterior$Fmsy,
    soi = tail(pdsi_soi, n = 1),
    Bt_BMSY = as.matrix(Bt_BMSY)
  )
  bassB_fmsy <- adjust_projection_jabba(
    FMSY = ss_case0$refpts_posterior$Fmsy,
    soi = tail(bassB_soi, n = 1),
    Bt_BMSY = as.matrix(Bt_BMSY)
  )
  effort_fmsy <- adjust_projection_jabba(
    FMSY = ss_case0$refpts_posterior$Fmsy,
    soi = tail(effort_soi, n = 1),
    Bt_BMSY = as.matrix(Bt_BMSY)
  )
  cpue_fmsy <- adjust_projection_jabba(
    FMSY = ss_case0$refpts_posterior$Fmsy,
    soi = tail(cpue_soi, n = 1),
    Bt_BMSY = as.matrix(Bt_BMSY)
  )
  
  if (projection_year_id == 1){
    fmsy_data <- data.frame(
      iter = 1:length(amo_fmsy),
      projection_year_id = projection_year[projection_year_id],
      JABBA = ss_case0$refpts_posterior$Fmsy,
      amo = amo_fmsy,
      pdsi = pdsi_fmsy,
      bassB = bassB_fmsy,
      effort = effort_fmsy, 
      cpue = cpue_fmsy
    )
  } else {
    fmsy_data <- rbind(
      fmsy_data,
      data.frame(
        iter = 1:length(amo_fmsy),
        projection_year_id = projection_year[projection_year_id],
        JABBA = ss_case0$refpts_posterior$Fmsy,
        amo = amo_fmsy,
        pdsi = pdsi_fmsy,
        bassB = bassB_fmsy,
        effort = effort_fmsy, 
        cpue = cpue_fmsy
      )
    )
  }
  
  
}

# Linear regression figure
lm_data <- rbind(lm_data_om, lm_data_em)
jpeg(filename = file.path(figure_path, paste0("linear_regression", terminal_year, ".jpeg")), width = 200, height = 100, units = "mm", res = 1800)
ggplot(lm_data, aes(x = Index_Value, y = Menhaden_Biomass, color = model)) +
  geom_point() +
  geom_smooth(method = lm) +
  facet_wrap(~Variable, scales = "free_x") +
  labs(
    title = "Linear regression analysis",
    x = "Indicator Value",
    y = "Menhaden spawning biomass (mt)"
  ) +
  theme_bw()
dev.off()

soi_data_melt <- reshape2::melt(
  soi_data,
  id = c("year", "projection_year_id")
)
soi_data_melt$projection_year_id <- as.factor(projection_year)

# SOI figure
jpeg(filename = file.path(figure_path, paste0("soi", terminal_year, ".jpeg")), width = 200, height = 100, units = "mm", res = 1800)
ggplot(soi_data_melt[soi_data_melt$projection_year_id==projection_year[1], ], aes(x = year, y = value)) +
  geom_line() +
  facet_wrap(~variable, scales = "free_y") +
  labs(
    title = "Status of Indicators",
    x = "Year",
    y = "Status of Indicators"
  ) +
  theme_bw()
dev.off()

fmsy_data_melt <- reshape2::melt(
  fmsy_data,
  id = c("iter", "projection_year_id")
)
fmsy_data_melt$projection_year_id <- as.factor(fmsy_data_melt$projection_year_id)

# ggplot(fmsy_data_melt, aes(x = iter, y = value, color = projection_year_id)) +
#   geom_line() +
#   facet_wrap(~variable) +
#   labs(
#     title = "FMSY",
#     x = "Iteration",
#     y = "FMSY"
#   ) +
#   theme_bw()

# FMSY figure
jpeg(filename = file.path(figure_path, paste0("FMSY", terminal_year, ".jpeg")), width = 200, height = 100, units = "mm", res = 1800)
ggplot(fmsy_data_melt, aes(x=variable, y = value, color = projection_year_id)) +
  geom_boxplot()+
  labs(
    title = "FMSY",
    x = "",
    y = "FMSY"
  ) +
  theme_bw()
dev.off()

fmsy_data_melt_median <- aggregate(value ~ projection_year_id+variable, data = fmsy_data_melt, median)
fmsy_data_melt_median$projection_year_id <- as.numeric(as.character(fmsy_data_melt_median$projection_year_id))

jpeg(filename = file.path(figure_path, paste0("median_fmsy", terminal_year, ".jpeg")), width = 200, height = 100, units = "mm", res = 1800)
ggplot(fmsy_data_melt_median, aes(x = projection_year_id, y = value, color = variable)) +
  geom_line() +
  labs(
    title = "Median FMSY from JABBA and adjusted FMSY",
    x = "",
    y = "FMSY"
  ) +
  theme_bw()
dev.off()


projection_f <- fmsy_data_melt_median
# projection_f[, paste0("f", sa_data$biodata$ages)] <- NA
fishery_sel <- IFA4EBFM::logistic(
    pattern = "double_logistic",
    x = sa_data$biodata$ages,
    slope_asc = 3.1,
    location_asc = 1.8,
    slope_desc = 0.88,
    location_desc = 0.01
) 


for (i in 1:nrow(projection_f)) {
    projection_f[i, paste0("fmsy_", sa_data$biodata$ages)] <- projection_f$value[i] * fishery_sel
    projection_f[i, paste0("fmsy0.5_", sa_data$biodata$ages)] <- projection_f$value[i] * fishery_sel*0.5
    projection_f[i, paste0("fmsy0.25_", sa_data$biodata$ages)] <- projection_f$value[i] * fishery_sel*0.25
}

write.csv(fmsy_data_melt, here::here("data", "data_moderate", "fmsy_data.csv"))
write.csv(projection_f, here::here("data", "data_moderate", "fmsy_median_data.csv"))

```

- Slope values from linear regression models

```{r, echo=FALSE, message=FALSE, warning=FALSE, fig.width=12, fig.height=10}
knitr::kable(lm_slope)
```

#### Adjust catch input and project biomass based on FMSY*B
```{r, results='hide', echo=FALSE, message=FALSE, warning=FALSE, fig.width=12, fig.height=10}

B2013 <- ss_case0$posteriors$BtoBmsy[, (length(model_year) + 1)] * ss_case0$posteriors$SBmsy[,1]

variable <- unique(fmsy_data_melt$variable)

Bt <- rep(B2013, times= length(variable))

tac_data_melt <- fmsy_data_melt
names(tac_data_melt)[names(tac_data_melt) == "value"] <- "fmsy"
tac_data_melt$tac <- adjust_projection_catcheco_jabba(fmsy_melted = fmsy_data_melt, Bt = Bt)
tac_data_melt_median <- aggregate(tac ~ variable, data = tac_data_melt, median)

ss_case0_input$jagsdata$nTAC <- length(variable)
ss_case0_input$jagsdata$TAC <- matrix(rep(tac_data_melt_median$tac, times =5), 
                                      ncol = length(variable), byrow = T)
ss_case0_input$settings$TAC.implementation <- projection_year[1]
ss_case0 <- JABBA::fit_jabba(
  jbinput = ss_case0_input,
  save.jabba = TRUE,
  save.all = TRUE,
  save.prj = TRUE,
  output.dir = output_dir,
  save.csvs = T
)

biomass_projection <- ss_case0$prj_posterior
biomass_projection <- biomass_projection[(biomass_projection$year %in% projection_year), ]
Bmsy <- rep(ss_case0$posteriors$SBmsy[, 1], 
            times= length(unique(biomass_projection$tac))*length(unique(biomass_projection$year)))
biomass_projection$biomass <- biomass_projection$stock*Bmsy
biomass_projection$year <- as.factor(biomass_projection$year)
biomass_projection$tac <- as.factor(biomass_projection$tac)

biomass_projection <- merge(biomass_projection, tac_data_melt_median, by="tac")

jpeg(filename = file.path(figure_path, paste0("biomass_projection", terminal_year, ".jpeg")), width = 200, height = 100, units = "mm", res = 1800)
ggplot(biomass_projection, aes(x=year, y = biomass, color = year)) +
  geom_boxplot()+
  facet_wrap(~variable)+
  labs(
    title = " Estimated biomass based on catch=FMSY*B",
    x = "",
    y = "Biomass"
  ) +
  theme_bw()
dev.off()
```

```{r, echo=FALSE, message=FALSE, warning=FALSE, fig.width=12, fig.height=10}
# om_K <- 5.12023947*1000000 # max b
# om_bmsy <- c(
#   4.553527815, 4.228060737, 2.657752888, 3.039367286, 1.700737553, 2.625573544, 1.885360091
# )*1000000
# om_fmsy <- c(
#   0.092402786, 0.344614655, 1.101458788, 0.722086847,
#   1.569404483, 0.988227546, 0.72328496
# )

om_K <- B0_F0_menhaden
om_bmsy <- compensation_msy$BMSY
om_fmsy <- compensation_msy$FMSY

cases <- c("ecosim_data_moderate_jabba_highF",
           "ecosim_data_moderate_amo_highF",
           "ecosim_data_moderate_pdsi_highF",
           "ecosim_data_moderate_bassB_highF",
           "ecosim_data_moderate_effort_highF")

evaluation_table <- data.frame(
  Average_Catch = NA,
  B2017_K = NA, 
  B2017_BMSY = NA,
  Average_Biomass = NA,
  Bonanza_Period = NA,
  Collapse_Period = NA
)

for (i in seq_along(cases)){
  file_path <- here::here("data", "ewe", "7ages_newsim_final", "ewe7ages_ecosim_om2_projections", cases[i])
  # Load EwE biomass data
  functional_groups <- c(
    "StripedBass0",
    "StripedBass2_5",
    "StripedBass6",
    "AtlanticMenhaden0",
    "AtlanticMenhaden1",
    "AtlanticMenhaden2",
    "AtlanticMenhaden3",
    "AtlanticMenhaden4",
    "AtlanticMenhaden5",
    "AtlanticMenhaden6",
    "SpinyDogfish",
    "BluefishJuvenile",
    "BluefishAdult",
    "WeakfishJuvenile",
    "WeakfishAdult",
    "AtlanticHerring0_1",
    "AtlanticHerring2",
    "Anchovies",
    "Benthos",
    "Zooplankton",
    "Phytoplankton",
    "Detritus"
  )
  
  age_name <- paste0("AtlanticMenhaden", 0:6)
  
  biomass <- read_ewe_output(
    file_path = file_path,
    file_names = "biomass_monthly.csv",
    skip_nrows = 8,
    plot = FALSE,
    figure_titles = NULL,
    functional_groups = functional_groups,
    figure_colors = NULL
  )
  
  time_id <- seq(1, nrow(biomass[[1]]), by = 12)
  biomass_ewe <- apply(biomass[[1]][, age_name], 1, sum) * 1000000
  
  catch <- read_ewe_output(
    file_path = file_path,
    file_names = "catch_monthly.csv",
    skip_nrows = 8,
    plot = FALSE,
    figure_titles = NULL,
    functional_groups = functional_groups,
    figure_colors = NULL
  )
  catch_ewe <- apply(catch[[1]][, age_name], 1, sum) * 1000000
  
  evaluation_table[i, ] <- c(
    mean(tail(catch_ewe[time_id], n=length(projection_year))),
    tail(biomass_ewe[time_id], n=1)/om_K,
    mean(tail(biomass_ewe[time_id], n=1)/om_bmsy),
    mean(tail(biomass_ewe[time_id], n=length(projection_year))),
    sum((tail(biomass_ewe[time_id], n=length(projection_year))/om_K)>0.8),
    sum((tail(biomass_ewe[time_id], n=length(projection_year))/om_K)<0.2)
  )
}

case_names <- levels(biomass_projection$variable)
for (i in 1:length(levels(biomass_projection$tac))) {
  projected_biomass <- aggregate(biomass ~ year + tac, data = biomass_projection, median)
  bk <- aggregate(bk ~ year + tac + variable, data = biomass_projection, median)
  
  evaluation_table <- rbind(
    evaluation_table,
    c(
      as.numeric(as.character(unique(biomass_projection$tac[biomass_projection$variable == case_names[i]]))),
      # median(biomass_projection$harvest[biomass_projection$year == tail(projection_year, n = 1) &
      #   biomass_projection$tac == unique(biomass_projection$tac[biomass_projection$variable == case_names[i]])]),
      median(biomass_projection$bk[biomass_projection$year == tail(projection_year, n = 1) &
        biomass_projection$tac == unique(biomass_projection$tac[biomass_projection$variable == case_names[i]])]),
      median(biomass_projection$stock[biomass_projection$year == tail(projection_year, n = 1) &
        biomass_projection$tac == unique(biomass_projection$tac[biomass_projection$variable == case_names[i]])]),
      mean(projected_biomass$biomass[projected_biomass$tac == unique(biomass_projection$tac[biomass_projection$variable == case_names[i]])]),
      sum(bk$bk[bk$tac == unique(bk$tac[bk$variable == case_names[i]])] > 0.8),
      sum(bk$bk[bk$tac == unique(bk$tac[bk$variable == case_names[i]])] < 0.2)
      
    )
  )
}

evaluation_table <- rbind(
  apply(evaluation_table, 2, max),
  rep(0, ncol(evaluation_table)),
  evaluation_table
)

row.names(evaluation_table) <- c("Max", "Min",
                                 "JABBA OM", 
                                 "AMO OM",
                                 "PDSI OM",
                                 "Bass Biomass OM",
                                 "Effot OM",
                                 "JABBA Default",
                                 "JABBA+AMO",
                                 "JABBA+PDSI",
                                 "JABBA+Bass Biomass",
                                 "JABBA+Effort")

evaluation_table["Max", c("Bonanza_Period", "Collapse_Period")] <- c(length(projection_year), length(projection_year))
evaluation_table[c("Max", "Min"), c("Collapse_Period")] <- rev(evaluation_table[c("Max", "Min"), c("Collapse_Period")])


case_num <- 4
par(mfrow=c(2,3), mar=c(0,0,0,0))
colors <- brewer.pal(12,"Paired")[-c(6, 12)]

variable_label <- c("Mean Catch", "B2017/K", "B2017/BMSY", "Mean Biomass", "Bonanza \n Length", "Minimize \n Collapse \n Length")
for (i in 1:case_num){
  colors_border <- colors[c(1, 1+i, 6, 6+i)]
  radarchart(evaluation_table[c(1:3, 3+i, 8, 8+i), ],
             seg=5,
             pcol=colors_border, plwd=2, plty=1,
             axistype = 0, axislabcol="black", 
             paxislabels = round(evaluation_table["Max",],2), palcex = 1,
             vlabels = variable_label, vlcex = 0.9)
}

colors_border <- colors[c(6:10)]
radarchart(evaluation_table[c(1:2, 8:12),],
           seg=5,
           pcol=colors_border, plwd=2 , plty=1,
           axistype = 2, axislabcol="black", 
           paxislabels = round(evaluation_table["Max",],2), palcex = 1,
           vlabels = variable_label, vlcex = 0.9)

colors_border <- colors[1:10]
plot(1, type = "n", axes=FALSE, xlab="", ylab="")
legend("center", legend = rownames(evaluation_table[-c(1,2),]), bty = "n", pch=20 , col=colors_border, text.col = "black", cex=0.9, pt.cex=3)

```

#### Adjust catch input and project biomass based on 0.5*FMSY*B

```{r, results='hide', echo=FALSE, message=FALSE, warning=FALSE, fig.width=12, fig.height=10}
tac_data_melt_median <- aggregate(tac ~ variable, data = tac_data_melt, median)
tac_data_melt_median$tac <- tac_data_melt_median$tac/2

ss_case0_input$jagsdata$nTAC <- length(variable)
ss_case0_input$jagsdata$TAC <- matrix(rep(tac_data_melt_median$tac, times =5), 
                                      ncol = length(variable), byrow = T)
ss_case0_input$settings$TAC.implementation <- 2013
ss_case0 <- JABBA::fit_jabba(
  jbinput = ss_case0_input,
  save.jabba = TRUE,
  save.all = TRUE,
  save.prj = TRUE,
  output.dir = output_dir,
  save.csvs = T
)

biomass_projection <- ss_case0$prj_posterior
biomass_projection <- biomass_projection[(biomass_projection$year %in% projection_year), ]
Bmsy <- rep(ss_case0$posteriors$SBmsy[, 1], 
            times= length(unique(biomass_projection$tac))*length(unique(biomass_projection$year)))
biomass_projection$biomass <- biomass_projection$stock*Bmsy
biomass_projection$year <- as.factor(biomass_projection$year)
biomass_projection$tac <- as.factor(biomass_projection$tac)

biomass_projection <- merge(biomass_projection, tac_data_melt_median, by="tac")

ggplot(biomass_projection, aes(x=year, y = biomass, color = year)) +
  geom_boxplot()+
  facet_wrap(~variable)+
  labs(
    title = " Estimated biomass based on catch=0.5*FMSY*B",
    x = "",
    y = "Biomass"
  ) +
  theme_bw()
```

```{r, echo=FALSE, message=FALSE, warning=FALSE, fig.width=12, fig.height=10}
om_K <- 5.12023947*1000000 # max b
om_bmsy <- c(
  4.553527815, 4.228060737, 2.657752888, 3.039367286, 1.700737553, 2.625573544, 1.885360091
)*1000000
om_fmsy <- c(
  0.092402786, 0.344614655, 1.101458788, 0.722086847,
  1.569404483, 0.988227546, 0.72328496
)

cases <- c("ecosim_data_moderate_jabba_mediumF",
           "ecosim_data_moderate_amo_mediumF",
           "ecosim_data_moderate_pdsi_mediumF",
           "ecosim_data_moderate_bassB_mediumF",
           "ecosim_data_moderate_effort_mediumF")

evaluation_table <- data.frame(
  Average_Catch = NA,
  B2017_K = NA, 
  B2017_BMSY = NA,
  Average_Biomass = NA,
  Bonanza_Period = NA,
  Collapse_Period = NA
)

for (i in seq_along(cases)){
  file_path <- here::here("data", "ewe", "7ages_newsim_final", "ewe7ages_ecosim_om2_projections", cases[i])
  # Load EwE biomass data
  functional_groups <- c(
    "StripedBass0",
    "StripedBass2_5",
    "StripedBass6",
    "AtlanticMenhaden0",
    "AtlanticMenhaden1",
    "AtlanticMenhaden2",
    "AtlanticMenhaden3",
    "AtlanticMenhaden4",
    "AtlanticMenhaden5",
    "AtlanticMenhaden6",
    "SpinyDogfish",
    "BluefishJuvenile",
    "BluefishAdult",
    "WeakfishJuvenile",
    "WeakfishAdult",
    "AtlanticHerring0_1",
    "AtlanticHerring2",
    "Anchovies",
    "Benthos",
    "Zooplankton",
    "Phytoplankton",
    "Detritus"
  )
  
  age_name <- paste0("AtlanticMenhaden", 0:6)
  
  biomass <- read_ewe_output(
    file_path = file_path,
    file_names = "biomass_monthly.csv",
    skip_nrows = 8,
    plot = FALSE,
    figure_titles = NULL,
    functional_groups = functional_groups,
    figure_colors = NULL
  )
  
  time_id <- seq(1, nrow(biomass[[1]]), by = 12)
  biomass_ewe <- apply(biomass[[1]][, age_name], 1, sum) * 1000000
  
  catch <- read_ewe_output(
    file_path = file_path,
    file_names = "catch_monthly.csv",
    skip_nrows = 8,
    plot = FALSE,
    figure_titles = NULL,
    functional_groups = functional_groups,
    figure_colors = NULL
  )
  catch_ewe <- apply(catch[[1]][, age_name], 1, sum) * 1000000
  
  evaluation_table[i, ] <- c(
    mean(tail(catch_ewe[time_id], n=length(projection_year))),
    tail(biomass_ewe[time_id], n=1)/om_K,
    mean(tail(biomass_ewe[time_id], n=1)/om_bmsy),
    mean(tail(biomass_ewe[time_id], n=length(projection_year))),
    sum((tail(biomass_ewe[time_id], n=length(projection_year))/om_K)>0.8),
    sum((tail(biomass_ewe[time_id], n=length(projection_year))/om_K)<0.2)
  )
}

case_names <- levels(biomass_projection$variable)
for (i in 1:length(levels(biomass_projection$tac))) {
  projected_biomass <- aggregate(biomass ~ year + tac, data = biomass_projection, median)
  bk <- aggregate(bk ~ year + tac + variable, data = biomass_projection, median)
  
  evaluation_table <- rbind(
    evaluation_table,
    c(
      as.numeric(as.character(unique(biomass_projection$tac[biomass_projection$variable == case_names[i]]))),
      # median(biomass_projection$harvest[biomass_projection$year == tail(projection_year, n = 1) &
      #   biomass_projection$tac == unique(biomass_projection$tac[biomass_projection$variable == case_names[i]])]),
      median(biomass_projection$bk[biomass_projection$year == tail(projection_year, n = 1) &
                                     biomass_projection$tac == unique(biomass_projection$tac[biomass_projection$variable == case_names[i]])]),
      median(biomass_projection$stock[biomass_projection$year == tail(projection_year, n = 1) &
                                        biomass_projection$tac == unique(biomass_projection$tac[biomass_projection$variable == case_names[i]])]),
      mean(projected_biomass$biomass[projected_biomass$tac == unique(biomass_projection$tac[biomass_projection$variable == case_names[i]])]),
      sum(bk$bk[bk$tac == unique(bk$tac[bk$variable == case_names[i]])] > 0.8),
      sum(bk$bk[bk$tac == unique(bk$tac[bk$variable == case_names[i]])] < 0.2)
      
    )
  )
}

evaluation_table <- rbind(
  apply(evaluation_table, 2, max),
  rep(0, ncol(evaluation_table)),
  evaluation_table
)

row.names(evaluation_table) <- c("Max", "Min",
                                 "JABBA OM", 
                                 "AMO OM",
                                 "PDSI OM",
                                 "Bass Biomass OM",
                                 "Effot OM",
                                 "JABBA Default",
                                 "JABBA+AMO",
                                 "JABBA+PDSI",
                                 "JABBA+Bass Biomass",
                                 "JABBA+Effort")

evaluation_table["Max", c("Bonanza_Period", "Collapse_Period")] <- c(length(projection_year), length(projection_year))
evaluation_table[c("Max", "Min"), c("Collapse_Period")] <- rev(evaluation_table[c("Max", "Min"), c("Collapse_Period")])


case_num <- 4
par(mfrow=c(2,3), mar=c(0,0,0,0))
colors <- brewer.pal(12,"Paired")[-c(6, 12)]

variable_label <- c("Mean Catch", "B2017/K", "B2017/BMSY", "Mean Biomass", "Bonanza \n Length", "Minimize \n Collapse \n Length")
for (i in 1:case_num){
  colors_border <- colors[c(1, 1+i, 6, 6+i)]
  radarchart(evaluation_table[c(1:3, 3+i, 8, 8+i), ],
             seg=5,
             pcol=colors_border, plwd=2, plty=1,
             axistype = 0, axislabcol="black", 
             paxislabels = round(evaluation_table["Max",],2), palcex = 1,
             vlabels = variable_label, vlcex = 0.9)
}

colors_border <- colors[c(6:10)]
radarchart(evaluation_table[c(1:2, 8:12),],
           seg=5,
           pcol=colors_border, plwd=2 , plty=1,
           axistype = 2, axislabcol="black", 
           paxislabels = round(evaluation_table["Max",],2), palcex = 1,
           vlabels = variable_label, vlcex = 0.9)

colors_border <- colors[1:10]
plot(1, type = "n", axes=FALSE, xlab="", ylab="")
legend("center", legend = rownames(evaluation_table[-c(1,2),]), bty = "n", pch=20 , col=colors_border, text.col = "black", cex=0.9, pt.cex=3)

```

#### Adjust catch input and project biomass based on 0.25*FMSY*B
```{r, results='hide', echo=FALSE, message=FALSE, warning=FALSE, fig.width=12, fig.height=10}
tac_data_melt_median <- aggregate(tac ~ variable, data = tac_data_melt, median)
tac_data_melt_median$tac <- tac_data_melt_median$tac*0.25

ss_case0_input$jagsdata$nTAC <- length(variable)
ss_case0_input$jagsdata$TAC <- matrix(rep(tac_data_melt_median$tac, times =5), 
                                      ncol = length(variable), byrow = T)
ss_case0_input$settings$TAC.implementation <- 2013
ss_case0 <- JABBA::fit_jabba(
  jbinput = ss_case0_input,
  save.jabba = TRUE,
  save.all = TRUE,
  save.prj = TRUE,
  output.dir = output_dir,
  save.csvs = T
)

biomass_projection <- ss_case0$prj_posterior
biomass_projection <- biomass_projection[(biomass_projection$year %in% projection_year), ]
Bmsy <- rep(ss_case0$posteriors$SBmsy[, 1], 
            times= length(unique(biomass_projection$tac))*length(unique(biomass_projection$year)))
biomass_projection$biomass <- biomass_projection$stock*Bmsy
biomass_projection$year <- as.factor(biomass_projection$year)
biomass_projection$tac <- as.factor(biomass_projection$tac)

biomass_projection <- merge(biomass_projection, tac_data_melt_median, by="tac")

ggplot(biomass_projection, aes(x=year, y = biomass, color = year)) +
  geom_boxplot()+
  facet_wrap(~variable)+
  labs(
    title = " Estimated biomass based on catch=0.25*FMSY*B",
    x = "",
    y = "Biomass"
  ) +
  theme_bw()
```

```{r, echo=FALSE, message=FALSE, warning=FALSE, fig.width=12, fig.height=10}
om_K <- 5.12023947*1000000 # max b
om_bmsy <- c(
  4.553527815, 4.228060737, 2.657752888, 3.039367286, 1.700737553, 2.625573544, 1.885360091
)*1000000
om_fmsy <- c(
  0.092402786, 0.344614655, 1.101458788, 0.722086847,
  1.569404483, 0.988227546, 0.72328496
)

cases <- c("ecosim_data_moderate_jabba_lowF",
           "ecosim_data_moderate_amo_lowF",
           "ecosim_data_moderate_pdsi_lowF",
           "ecosim_data_moderate_bassB_lowF",
           "ecosim_data_moderate_effort_lowF")

evaluation_table <- data.frame(
  Average_Catch = NA,
  B2017_K = NA, 
  B2017_BMSY = NA,
  Average_Biomass = NA,
  Bonanza_Period = NA,
  Collapse_Period = NA
)

for (i in seq_along(cases)){
  file_path <- here::here("data", "ewe", "7ages_newsim_final", "ewe7ages_ecosim_om2_projections", cases[i])
  # Load EwE biomass data
  functional_groups <- c(
    "StripedBass0",
    "StripedBass2_5",
    "StripedBass6",
    "AtlanticMenhaden0",
    "AtlanticMenhaden1",
    "AtlanticMenhaden2",
    "AtlanticMenhaden3",
    "AtlanticMenhaden4",
    "AtlanticMenhaden5",
    "AtlanticMenhaden6",
    "SpinyDogfish",
    "BluefishJuvenile",
    "BluefishAdult",
    "WeakfishJuvenile",
    "WeakfishAdult",
    "AtlanticHerring0_1",
    "AtlanticHerring2",
    "Anchovies",
    "Benthos",
    "Zooplankton",
    "Phytoplankton",
    "Detritus"
  )
  
  age_name <- paste0("AtlanticMenhaden", 0:6)
  
  biomass <- read_ewe_output(
    file_path = file_path,
    file_names = "biomass_monthly.csv",
    skip_nrows = 8,
    plot = FALSE,
    figure_titles = NULL,
    functional_groups = functional_groups,
    figure_colors = NULL
  )
  
  time_id <- seq(1, nrow(biomass[[1]]), by = 12)
  biomass_ewe <- apply(biomass[[1]][, age_name], 1, sum) * 1000000
  
  catch <- read_ewe_output(
    file_path = file_path,
    file_names = "catch_monthly.csv",
    skip_nrows = 8,
    plot = FALSE,
    figure_titles = NULL,
    functional_groups = functional_groups,
    figure_colors = NULL
  )
  catch_ewe <- apply(catch[[1]][, age_name], 1, sum) * 1000000
  
  evaluation_table[i, ] <- c(
    mean(tail(catch_ewe[time_id], n=length(projection_year))),
    tail(biomass_ewe[time_id], n=1)/om_K,
    mean(tail(biomass_ewe[time_id], n=1)/om_bmsy),
    mean(tail(biomass_ewe[time_id], n=length(projection_year))),
    sum((tail(biomass_ewe[time_id], n=length(projection_year))/om_K)>0.8),
    sum((tail(biomass_ewe[time_id], n=length(projection_year))/om_K)<0.2)
  )
}

case_names <- levels(biomass_projection$variable)
for (i in 1:length(levels(biomass_projection$tac))) {
  projected_biomass <- aggregate(biomass ~ year + tac, data = biomass_projection, median)
  bk <- aggregate(bk ~ year + tac + variable, data = biomass_projection, median)
  
  evaluation_table <- rbind(
    evaluation_table,
    c(
      as.numeric(as.character(unique(biomass_projection$tac[biomass_projection$variable == case_names[i]]))),
      # median(biomass_projection$harvest[biomass_projection$year == tail(projection_year, n = 1) &
      #   biomass_projection$tac == unique(biomass_projection$tac[biomass_projection$variable == case_names[i]])]),
      median(biomass_projection$bk[biomass_projection$year == tail(projection_year, n = 1) &
                                     biomass_projection$tac == unique(biomass_projection$tac[biomass_projection$variable == case_names[i]])]),
      median(biomass_projection$stock[biomass_projection$year == tail(projection_year, n = 1) &
                                        biomass_projection$tac == unique(biomass_projection$tac[biomass_projection$variable == case_names[i]])]),
      mean(projected_biomass$biomass[projected_biomass$tac == unique(biomass_projection$tac[biomass_projection$variable == case_names[i]])]),
      sum(bk$bk[bk$tac == unique(bk$tac[bk$variable == case_names[i]])] > 0.8),
      sum(bk$bk[bk$tac == unique(bk$tac[bk$variable == case_names[i]])] < 0.2)
      
    )
  )
}

evaluation_table <- rbind(
  apply(evaluation_table, 2, max),
  rep(0, ncol(evaluation_table)),
  evaluation_table
)

row.names(evaluation_table) <- c("Max", "Min",
                                 "JABBA OM", 
                                 "AMO OM",
                                 "PDSI OM",
                                 "Bass Biomass OM",
                                 "Effot OM",
                                 "JABBA Default",
                                 "JABBA+AMO",
                                 "JABBA+PDSI",
                                 "JABBA+Bass Biomass",
                                 "JABBA+Effort")

evaluation_table["Max", c("Bonanza_Period", "Collapse_Period")] <- c(length(projection_year), length(projection_year))
evaluation_table[c("Max", "Min"), c("Collapse_Period")] <- rev(evaluation_table[c("Max", "Min"), c("Collapse_Period")])


case_num <- 4
par(mfrow=c(2,3), mar=c(0,0,0,0))
colors <- brewer.pal(12,"Paired")[-c(6, 12)]

variable_label <- c("Mean Catch", "B2017/K", "B2017/BMSY", "Mean Biomass", "Bonanza \n Length", "Minimize \n Collapse \n Length")
for (i in 1:case_num){
  colors_border <- colors[c(1, 1+i, 6, 6+i)]
  radarchart(evaluation_table[c(1:3, 3+i, 8, 8+i), ],
             seg=5,
             pcol=colors_border, plwd=2, plty=1,
             axistype = 0, axislabcol="black", 
             paxislabels = round(evaluation_table["Max",],2), palcex = 1,
             vlabels = variable_label, vlcex = 0.9)
}

colors_border <- colors[c(6:10)]
radarchart(evaluation_table[c(1:2, 8:12),],
           seg=5,
           pcol=colors_border, plwd=2 , plty=1,
           axistype = 2, axislabcol="black", 
           paxislabels = round(evaluation_table["Max",],2), palcex = 1,
           vlabels = variable_label, vlcex = 0.9)

colors_border <- colors[1:10]
plot(1, type = "n", axes=FALSE, xlab="", ylab="")
legend("center", legend = rownames(evaluation_table[-c(1,2),]), bty = "n", pch=20 , col=colors_border, text.col = "black", cex=0.9, pt.cex=3)

```
